Refactor ScoreLike to it's own file
