+++ src/main/scala/com/github/lstephen/ootp/ai/selection/Score.scala
@@ -1,7 +1,5 @@
 package com.github.lstephen.ootp.ai.selection
 
-import com.github.lstephen.ootp.ai.algebra._
-
 import spire.algebra.{ CMonoid, Order }
 import spire.math.Real
 
@@ -32,17 +30,4 @@ object Score {
 
   implicit val Alg: ScoreAlgebra = new ScoreAlgebra
 }
-
-trait ScoreLike {
-  def total: Score
-}
-
-object ScoreLike {
-  implicit def ord[A <: ScoreLike]: Order[A] = Order.by(_.total)
-
-  implicit class ScoreLikeSeq[A <: ScoreLike](xs: Seq[A]) {
-    def total: Score = xs.map(_.total).mconcat
-  }
-}
-
